+++ app/src/main/java/fr/esimed/searchcompany/data/model/Company.kt
@@ -5,7 +5,7 @@ import androidx.room.Index
 import androidx.room.PrimaryKey
 import java.io.Serializable
 @Entity(indices=[Index(value = ["id"],unique = true)])
-class Company(@PrimaryKey(autoGenerate = true)var id:Int?,var libelle:String, var depart:Int, var adress:String, var activity:String, var siret:String):Serializable {
+class Company(@PrimaryKey(autoGenerate = true)var id:Long?,var libelle:String, var depart:Int, var adress:String, var activity:String, var siret:String):Serializable {
 
     override fun equals(other: Any?): Boolean {
         if(this===other)return true
